#1835 Adjusted client implementation of sse handling and registration. Also Enhanced file processing behavior to use single evnt source for all file operations (uploading).
